Created an initial wrapper.
It will accept an input file containing some boxes and save them to an
internal array list.

The next step will be to create a bunch of genes (hence Bio) and try them out.
The genes will be a collection of boxes in different rotations and they will
be ranked according to their height.  The best ones will live to breed!